puntoverde/requirements.txt
asgiref==3.5.2
Django==4.1
django-crispy-forms==1.14.0
django-dotenv==1.4.2
django-environ==0.9.0
mysqlclient==2.1.1
sqlparse==0.4.2
tzdata==2022.2
